Deserialize now sets references in an object.
Deserialize now sets references properly in an object. Because of the nature of comparing keys in IdentityHashMap, namely the use of the == operator which checks the addresses of strings, reference element strings will not find an object in the map when used as a key (as the keys are attribute strings from object elements). This will need to be fixed, as the workaround currently in place defeats the purpose of a hashmap.