Extract Method on deserializeObject()
Extracted method getObject(Element, Class) from deserializeObject to prevent a future code duplication bad code smell. getObject will be used to set values for arrays and collections, and therefore would of needed to be copied and modified slightly. Instead of setting fields in the if statements, it now returns either the object referenced or a wrapper object containing the primitive data or a string.