Fixed lookup in hashmap
Switched IdentityHashMap to HashMap as HashMap enforces object equality meaning 2 different Integer objects with the same value will be equal even if they are not the same reference. This solves inputs read from the file being not found as attribute id values and reference element values were different object references and would not work as keys.